fix: consistency between run-local command and the rest
